Move pages to a 'pages/' folder
