+++ src/pages/Auth/SignUp/index.vue
