+++ src/pages/Auth/index.vue
