+++ src/pages/Home/BookDetails/styles.scss
